ACTION-2047: Remove mappings for role="text".
Commented out the entry for the 'text' role in the role mapping table.
The ARIA Working Group agreed to move role="text" to ARIA 2.0. See:
https://lists.w3.org/Archives/Public/public-aria-admin/2016Apr/0030.html

Unfortunately, we have not yet branched for 1.1 so in order to "remove"
it for 1.1, it's commented out. That way, it's gone for those reading
the spec, and preserved until we're ready to properly move it.