Core-AAM: Update UIA mapping for aria-autocomplete=list, etc (#627)
Similarly to AX API (as I understand) and differently from MSAA/IA2/ATK,
UIA doesn't have a concept of information that conveys the fact that
autocomplete is present. So while the proposal discussed in github issue
#419 makes some sense, that would be very brittle heuristic and not
something that I would comfortably specify generically.

Moreover when contrasted against other APIs mappings, this seems like a
purely informational mapping that conveys presence of the autocomplete
list (or other types of autocompletes), and there is just no such thing
in UIA. If that changes, we should revisit and add it here.