Core-AAM: Remove erroneous statement about not exposing submenus in ATK
The statement that elements with role menu "are not exposed for a
submenu if there is a parent menu" describes the complete opposite
of what is expected by platform assistive technologies. The fact
that ATK and AT-SPI2 don't programmatically prevent menu items from
having menu item children doesn't make such a hierarchy correct.