Editorial: Split role mapping for textbox into two tables in Core AAM
macOS has entirely different role mappings based on whether the value
of aria-multiline is true or false. In addition, IA2 and ATK have
different states associated with each condition. Thus breaking these
into two separate mapping tables improves readability.

Fix typo in the IA2 states (missing the "STATE").

Remove mention of STATE_SINGLE_LINE for searchbox for ATK: It is
redundant to the content of the state and property mapping tables
for aria-multiline, and is not mentioned for the other platforms.