UIA mapping for aria-placeholder
In the last release we made a change to map placeholder attribute and aria-placeholder to the UIA HelpText property (https://msdn.microsoft.com/en-us/library/dd757479(v=vs.85).aspx). This change reflects that.

Note: AriaProperties gets automatically populated with ALL aria-* properties, which is very easy way to "map" something, when there is no better alternative. Now that we found alternative for placeholder - we can actually specify it properly.