Update ATK mappings for alert and alertdialog roles
* alertdialog: ATK_ROLE_ALERT rather than ATK_ROLE_DIALOG
* alert: ATK_ROLE_NOTIFICATION rather than ATK_ROLE_ALERT

In ARIA, an alertdialog is a type of dialog; an alert is not a type of
dialog but rather a live region.

In ATK an alert is a type of dialog which should be used for things
like warning dialogs. Thus this platform role is well suited for the
ARIA alertdialog role, but ill suited for the ARIA alert role.

Resolves #68 and also resolves #69.