ATK: Update mappings for table row and column properties
For the following properties:
* aria-colindex
* aria-colspan
* aria-colcount
* aria-rowindex
* aria-rowspan
* aria-rowcount

The object attribute should contain the author-provided value, but
AtkTable should reflect the actual value. If the values exposed by
AtkTable don't reflect the physical table, AT-provided table navigation
can be buggy and non-performant.

Firefox and Chromium already have implemented AtkTable as described
above, and Orca already uses that support. Thus this change is bringing
the spec in alignment with reality.