remove reference to whitespace (#135)
* remove reference to whitespace

resolves #128

Per https://www.w3.org/2022/08/11-aria-minutes#t05, the WG resolved that the empty string should cause the attribute to be ignored, but not whitespace.  So, reference to whitespace can be removed from this spec.  However, referring to whitespace will remain in the ARIA spec and THERE it will link to the ascii-whitespace definition where applicable.   Additionally, rather than a user agent requirement, it will become an authors requirement on not specifying whitespace as the value of an aria-roledescription

* Update index.html